Add files via upload
A Wikipedia link scraper I made to find "degrees of separation to Kevin Bacon (on Wikipedia)". However, I decided to publish only the first level of link-scraping because of computing time limitations (time grows exponentially with each level). The user can paste in a starting Wikipedia link, and this will scrape all the links on that page for mentions of Kevin Bacon.